requirements.txt
allure-pytest==2.12.0
pydantic==2.5.1
pydantic-settings==2.1.0
pytest==7.2.1
python-dotenv==0.21.1
pytest-rerunfailures==13.0
selenium==4.8.0
selenium-wire==5.1.0
webdriver-manager==4.0.1
